refactoring lane processing
