+++ src/ImageProcessing/LaneDetect/LaneDetector.py
@@ -23,20 +23,6 @@ class LaneDetector:
                 (int(self.width * 0.2), self.height // 2 - int(self.height * 0.05))
             ]], np.int32)
         
-        self.stopReg = np.array([[
-                (self.width*0.75, self.height*0.68),
-                (self.width*0.25, self.height*0.68),
-                (self.width*0.25, self.height*0.83),
-                (self.width*0.75, self.height*0.83)
-            ]], np.int32)
-        
-        self.interStopReg = np.array([[
-                (self.width*0.68, self.height*0.40),
-                (self.width*0.32, self.height*0.40),
-                (self.width*0.25, self.height*0.68),
-                (self.width*0.75, self.height*0.68)
-            ]], np.int32)
-        
         self.squareLeft = (100, 180, 190, 220)
         self.squareRight = (320, 180, 410, 220)
 
@@ -121,42 +107,6 @@ class LaneDetector:
     
     def distance(self, point1, point2):
         return math.sqrt((point1[0] - point2[0])**2 + (point1[1] - point2[1])**2)
-
-    def detectIntersection(self, lines) -> bool:
-        if lines is None:
-            return False, []
-        
-        lines2 = []
-        for line in lines:
-            for x1, y1, x2, y2 in line:
-                slope = (y2 - y1) / (x2 - x1) if x2 != x1 else 0
-                distance = np.sqrt((x2 - x1)**2 + (y2 - y1)**2)
-
-                if slope < 0.2 and slope > -0.2:
-                    if distance > 50:
-                        lines2.append([(x1, y1), (x2, y2)])
-
-        if(len(lines2) >= 2):
-            return True, lines2
-
-        return False, []
-    def gamma_correction(self, image: np.ndarray, gamma: float) -> np.ndarray:
-        image_normalized = image  / 255.0
-        corrected = np.power(image_normalized, gamma)
-        corrected = np.uint8(corrected * 255)
-        return corrected
-
-    def region_of_interest2(self, img):
-        mask = np.zeros_like(img, dtype=np.uint8)
-        cv2.fillPoly(mask, self.stopReg, (555, 255, 255)) # Mask now matches BGR format
-        masked_image = cv2.bitwise_and(img, mask)
-        return masked_image
-    
-    def region_of_interest3(self, img):
-        mask = np.zeros_like(img, dtype=np.uint8)
-        cv2.fillPoly(mask, self.interStopReg, (255, 255, 255)) # Mask now matches BGR format
-        masked_image = cv2.bitwise_and(img, mask)
-        return masked_image
     
     def drawRectangle(self, frame, square):
         x1, y1, x2, y2 = square
@@ -165,29 +115,14 @@ class LaneDetector:
     def detect_lines(self, img, tres = 30):
         return cv2.HoughLinesP(img, rho=1, theta=np.pi / 180, threshold=tres, minLineLength=15, maxLineGap=5)
     
-    def process_frame(self, frame: np.ndarray):
+    def process_frame(self, frame: np.ndarray, edges):
         angle_degrees: float = 0.0
 
-        gray = cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY)
-        gama = self.gamma_correction(gray, 12)
-        blurred = cv2.GaussianBlur(gama, [5, 5], 0)
-        edges = cv2.Canny(blurred, 50, 150)
-
-        roi2 = self.region_of_interest2(edges)
         roi = self.region_of_interest(edges)
-        roi3 = self.region_of_interest3(edges)
-
         lines = self.detect_lines(roi, 25)
-        lines2 = self.detect_lines(roi2, 10)
-        lines3 = self.detect_lines(roi3, 15)
-
-        intersection, linesX = self.detectIntersection(lines2)
-        intersectionA, linesY = self.detectIntersection(lines3)
-
         angle_degrees = float(self.calculate_steering_angle(lines, frame.shape[1], frame.shape[0]))
 
         if self.debugging:
-            # Draw lines on the frame
             if lines is not None:
                 for line in lines:
                     for x1, y1, x2, y2 in line:
@@ -195,31 +130,11 @@ class LaneDetector:
                         if slope < -0.4 or slope > 0.4:
                             cv2.line(frame, (x1, y1), (x2, y2), (0, 255, 0), 3)
             
-
             self.drawRectangle(frame, self.squareLeft)
             self.drawRectangle(frame, self.squareRight)
 
             points = self.roadReg.reshape((-1, 1, 2))
             cv2.polylines(frame, [points], isClosed=True, color=(0, 255, 0), thickness=2)
-            points = self.stopReg.reshape((-1, 1, 2))
-            cv2.polylines(frame, [points], isClosed=True, color=(255, 255, 0), thickness=2)
-            points = self.interStopReg.reshape((-1, 1, 2))
-            cv2.polylines(frame, [points], isClosed=True, color=(150, 255, 50), thickness=2)
 
-            if len(linesX) > 0:
-                for line in linesX:
-                    (x1, y1), (x2, y2) = line
-                    cv2.line(frame, (x1, y1), (x2, y2), (255, 0, 0), 3)
-            
-            if len(linesY) > 0:
-                for line in linesY:
-                    (x1, y1), (x2, y2) = line
-                    cv2.line(frame, (x1, y1), (x2, y2), (200, 50, 0), 3)
-        
-        if not self.pc:
-            self.strm.display(frame)
 
-        if intersectionA:
-            intersection = False
-
-        return frame, angle_degrees, intersection, intersectionA
+        return frame, angle_degrees
